+++ CHANGELOG.md
@@ -11,6 +11,7 @@ To see compatibility information, `xsvutils help compatibility`.
 - add `filter`, `where` subcommand on v2
 - change --o-diffable format
 - add some subcommands from v1 to v2
+- add `insmap` subcommand on v2
 
 ## version 0.2.6 (2018/04/09)
 - add support LTSV format on v2